Fix segfault when calling statvfs on 32-bit linux
On 32-bit linux, ruby's configure script detects that FILE_OFFSET_BITS=64 is
required to access the 64-bit filesystem interface and adds the following to
rbconfig.rb:

    CONFIG["DEFS"] = "-D_FILE_OFFSET_BITS=64"

The sys-filesystem gem was modified in 68d6cdb978d4d25b25d9eafa5e1e5b5642e88578
to detect that case and call the 64-bit version of `statvfs`. The `Statfs`
struct was also modified, but the similarly named `Statvfs` struct wasn't. So
the following segfaulted on 32-bit linux:

    $ ruby -rsys-filesystem -e "Sys::Filesystem.stat('/')"
    *** glibc detected *** /opt/puppetlabs/puppet/bin/ruby: double free or corruption (!prev): 0x094b1c58 ***